requirements.txt
adafruit-circuitpython-servokit
eventlet
python-socketio